+++ cmd/migrate/migrate_test.go
@@ -16,20 +16,28 @@ import (
 const defaultDuration = 1 * time.Minute
 
 func TestMigrateCommandRollbacks(t *testing.T) {
-	engines := []string{"postgres", "mysql"}
+	type EngineConfig struct {
+		Engine     string
+		MinVersion int64
+	}
+	engines := []EngineConfig{
+		{Engine: "postgres"},
+		{Engine: "mysql"},
+		{Engine: "sqlite", MinVersion: 5},
+	}
 
-	for _, engine := range engines {
-		t.Run(engine, func(t *testing.T) {
-			container, _, uri := util.MustBootstrapDatastore(t, engine)
+	for _, e := range engines {
+		t.Run(e.Engine, func(t *testing.T) {
+			container, _, uri := util.MustBootstrapDatastore(t, e.Engine)
 
 			// going from version 3 to 4 when migration #4 doesn't exist is a no-op
 			version := container.GetDatabaseSchemaVersion() + 1
 
 			migrateCommand := NewMigrateCommand()
 
-			for version >= 0 {
+			for version >= e.MinVersion {
 				t.Logf("migrating to version %d", version)
-				migrateCommand.SetArgs([]string{"--datastore-engine", engine, "--datastore-uri", uri, "--version", strconv.Itoa(int(version))})
+				migrateCommand.SetArgs([]string{"--datastore-engine", e.Engine, "--datastore-uri", uri, "--version", strconv.Itoa(int(version))})
 				err := migrateCommand.Execute()
 				require.NoError(t, err)
 				version--