+++ cmd/run/run_test.go
@@ -794,6 +794,9 @@ func TestServerMetricsReporting(t *testing.T) {
 	t.Run("postgres", func(t *testing.T) {
 		testServerMetricsReporting(t, "postgres")
 	})
+	t.Run("sqlite", func(t *testing.T) {
+		testServerMetricsReporting(t, "sqlite")
+	})
 }
 
 func testServerMetricsReporting(t *testing.T, engine string) {