+++ pkg/storage/record.go
@@ -17,7 +17,10 @@ type TupleRecord struct {
 	ObjectType       string
 	ObjectID         string
 	Relation         string
-	User             string
+	User             string // Deprecated: Use UserObjectType, UserObjectID & UserRelation instead.
+	UserObjectType   string
+	UserObjectID     string
+	UserRelation     string
 	ConditionName    string
 	ConditionContext *structpb.Struct
 	Ulid             string
@@ -26,11 +29,16 @@ type TupleRecord struct {
 
 // AsTuple converts a [TupleRecord] into a [*openfgav1.Tuple].
 func (t *TupleRecord) AsTuple() *openfgav1.Tuple {
+	user := t.User
+	if t.User == "" {
+		user = tupleutils.FromUserParts(t.UserObjectType, t.UserObjectID, t.UserRelation)
+	}
+
 	return &openfgav1.Tuple{
 		Key: tupleutils.NewTupleKeyWithCondition(
 			tupleutils.BuildObject(t.ObjectType, t.ObjectID),
 			t.Relation,
-			t.User,
+			user,
 			t.ConditionName,
 			t.ConditionContext,
 		),