+++ tests/check/check_test.go
@@ -9,6 +9,7 @@ import (
 	"net/http"
 	"strconv"
 	"testing"
+	"time"
 
 	"github.com/oklog/ulid/v2"
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
@@ -52,6 +53,8 @@ func testRunTestMatrix(t *testing.T, engine string, experimental bool) {
 		cfg.Datastore.Engine = engine
 		cfg.ListUsersDeadline = 0   // no deadline
 		cfg.ListObjectsDeadline = 0 // no deadline
+		// extend the timeout for the tests, coverage makes them slower
+		cfg.RequestTimeout = 10 * time.Second
 
 		tests.StartServer(t, cfg)
 
@@ -71,6 +74,10 @@ func TestCheckPostgres(t *testing.T) {
 
 func TestCheckMySQL(t *testing.T) {
 	testRunAll(t, "mysql")
+}
+
+func TestCheckSQLite(t *testing.T) {
+	testRunAll(t, "sqlite")
 }
 
 // TODO move elsewhere as this isn't asserting on just Check API logs.
@@ -326,6 +333,8 @@ func testRunAll(t *testing.T, engine string) {
 	cfg := config.MustDefaultConfig()
 	cfg.Log.Level = "error"
 	cfg.Datastore.Engine = engine
+	// extend the timeout for the tests, coverage makes them slower
+	cfg.RequestTimeout = 10 * time.Second
 
 	tests.StartServer(t, cfg)
 
@@ -344,6 +353,10 @@ func BenchmarkCheckPostgres(b *testing.B) {
 
 func BenchmarkCheckMySQL(b *testing.B) {
 	benchmarkAll(b, "mysql")
+}
+
+func BenchmarkCheckSQLite(b *testing.B) {
+	benchmarkAll(b, "sqlite")
 }
 
 func benchmarkAll(b *testing.B, engine string) {
@@ -392,6 +405,8 @@ func setupBenchmarkTest(b *testing.B, engine string) (openfgav1.OpenFGAServiceCl
 	cfg := config.MustDefaultConfig()
 	cfg.Log.Level = "none"
 	cfg.Datastore.Engine = engine
+	// extend the timeout for the tests, coverage makes them slower
+	cfg.RequestTimeout = 10 * time.Second
 
 	tests.StartServer(b, cfg)